enhancement: 키워드의 이슈 요약을 가져오는 모듈 분리
