+++ app/report/__init__.py
@@ -1,0 +1,1 @@
+from app.report.toady_issue_summary import get_today_issue_summary
