+++ app/request_external_api/__init__.py
@@ -1,5 +1,3 @@
-from app.request_external_api.headers import CLOVA_API_HEADERS
-from app.request_external_api.headers import CLOVA_SUMMARY_API_ENDPOINT
 from app.request_external_api.request_news import RequestNews
 from app.request_external_api.request_suggestions import RequestSuggestions
 from app.request_external_api.request_trend import RequestTrend
