--- app/request_external_api/headers.py
@@ -1,10 +1,0 @@
-import os
-
-
-CLOVA_API_HEADERS = {
-    "X-NCP-APIGW-API-KEY-ID": os.getenv('CLOVA_API_CLIENT_ID'),
-    "X-NCP-APIGW-API-KEY": os.getenv('CLOVA_API_CLIENT_SECRET'),
-    "Content-Type": "application/json"
-}
-
-CLOVA_SUMMARY_API_ENDPOINT = "https://naveropenapi.apigw.ntruss.com/text-summary/v1/summarize"
