+++ app/request_external_api/request_news.py
@@ -45,3 +45,10 @@ class RequestNews:
 #         return None
 #     except Exception as err:
 #         print(f"Other error occurred: {err}")  # 기타 에러 출력
+#CLOVA_API_HEADERS = {
+#     "X-NCP-APIGW-API-KEY-ID": os.getenv('CLOVA_API_CLIENT_ID'),
+#     "X-NCP-APIGW-API-KEY": os.getenv('CLOVA_API_CLIENT_SECRET'),
+#     "Content-Type": "application/json"
+# }
+#
+# CLOVA_SUMMARY_API_ENDPOINT = "https://naveropenapi.apigw.ntruss.com/text-summary/v1/summarize"
